Updated gathercontent-downloadfiles
read the list of files from the local files.json
change the name of the callback from done, to cb so that it doesn't
override async.
add a callback to the request, so that each file will call it once it
has downloaded.
when all the files are downloaded, call the done(); function